Split modular pipeline code into a package
